move get_uuid from a property to a utility method
This change will ensure that director is more accessible as a library.
